0.5.0

+++ common/src/constants.rs
@@ -1,5 +1,5 @@
 /// Refresh token length in ascii
-pub const REFRESH_TOKEN_LEN: usize = 256;
+pub const REFRESH_TOKEN_LEN: u64 = 256;
 /// Default duration of ID Token validity in minutes
 /// TODO: Clapで設定できるように
 pub const JWT_DURATION_MINS: usize = 30;
